+++ python/GafferSceneTest/__init__.py
@@ -168,6 +168,7 @@ from .SetQueryTest import SetQueryTest
 from .VisibleSetTest import VisibleSetTest
 from .VisibleSetDataTest import VisibleSetDataTest
 from .MeshSegmentsTest import MeshSegmentsTest
+from .ImageToPointsTest import ImageToPointsTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *